refactor: consolidate Vitest configuration and update Storybook stories
