+++ .storybook/preview.ts
@@ -1,12 +1,14 @@
-import type { Preview } from '@storybook/react'
-import '../app/globals.css'
+import type { Preview } from '@storybook/react';
+import '../app/globals.css';
 
 const preview: Preview = {
   parameters: {
+    react: { rsc: true },
+    nextjs: { appDirectory: true },
     controls: {
       matchers: {
-       color: /(background|color)$/i,
-       date: /Date$/i,
+        color: /(background|color)$/i,
+        date: /Date$/i,
       },
     },
   },
