+++ app/utils/sum.test.ts
@@ -1,6 +1,6 @@
-import { sum } from "./sum";
-import { expect, test } from "vitest";
+import { sum } from './sum';
+import { expect, test } from 'vitest';
 
-test("sum", () => {
+test('sum', () => {
   expect(sum(1, 2)).toBe(3);
 });
